requirements.txt
boto3==1.26.118
boto3==1.26.118
coverage==7.2.7
dj-rest_auth==4.0.1
django-anymail==9.1
django-cors-headers==4.2.0
django-debug-toolbar==4.0.0
django-filter==23.2
django-simple-history==3.3.0
djangorestframework==3.14.0,<3.15.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
drf-spectacular>=0.26.0,<0.27.0
Faker==18.7.0
gunicorn==20.1.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.30.0
sib-api-v3-sdk==7.6.0
tox==4.6.4
uvicorn==0.20.0
pandas==2.2.2